cloud_functions/register_model/requirements.txt
google-cloud-aiplatform==1.27.0 
google-cloud-storage==2.9.0 
functions-framework==3.2.0
